fix: Complete LOW priority items from external code review (OpenSpec #00038)
LOW priority fixes implemented:
- [3/24/30] QPointer for DashboardPanel (safe pointer tracking)
- [4] getWordCount() returns size_t instead of int
- [15] noexcept specifications for getters and simple functions
- [11] clearLayout utility used everywhere (DRY)
- [28] Command&& move overload for registerCommand
- [23] ZipSourceBuffer RAII wrapper for libzip operations

All HIGH, MEDIUM, and LOW priority items from external code review now complete.

+++ src/gui/panels/properties_panel.cpp
@@ -833,7 +833,7 @@ void PropertiesPanel::updateProjectStatistics() {
 
     // Get chapter count and word count
     size_t chapterCount = book.getChapterCount();
-    int wordCount = book.getWordCount();
+    size_t wordCount = book.getWordCount();
 
     m_projectChaptersLabel->setText(QString::number(chapterCount));
     m_projectWordsLabel->setText(QString::number(wordCount));
